backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
sqlparse==0.5.3
typing_extensions==4.15.0

# Production dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
PyMySQL==1.1.0
